revisit entity expansion for xml content
Make solid the 5 known character entities, hex digit entities, decimal digit entities
convert the expanded code points to utf-8 sequences
move more enum conversions into converter.h
switch attributes to load from xmlcollection
play with marker orientation
remove SVG_UNITS_XXX in favor of SVGLength::SVG_LENGTHTYPE

+++ app/appmain.cpp
@@ -90,7 +90,7 @@ unsigned int systemPpi = 192;   // starting pixel density
 
 // Stuff related to rate of displaying frames
 float fFrameRate = 1;
-float fInterval = 1000;
+double fInterval = 1000;
 double fNextMillis = 0;
 size_t fDroppedFrames = 0;
 uint64_t fFrameCount = 0;         // how many frames drawn so far
